+++ src/main/java/com/example/chuyendeweb/repository/LocationRepository.java
@@ -3,5 +3,5 @@ package com.example.chuyendeweb.repository;
 import com.example.chuyendeweb.entities.location;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface LocationRepository extends JpaRepository<location,Long> {
+public interface LocationRepository extends JpaRepository<location,Integer> {
 }
